Set permissions and ownership in postinst
Android's make_ext4fs, used to create the sparse ext4 filesystem
images which are written to Android devices via the fastboot protocol,
does not preserve file ownership. With the '-a' option, it sets
ownership via hardcoded uids, guids, filepaths in the make_ext4fs
source, but it won't preserve the ownership of the files it's
bundling. Yuck! Googled again.

As a workaround, set the ownership and permissions of the files in
this package via postinst.